增加了说明，并修改了一些bug，初步在archlinux 64位系统上编过

+++ api/helper.h
@@ -15,7 +15,7 @@
     do { typeof(a) __tmp = (a); (a) = (b); (b) = __tmp; } while (0)
 
 #ifndef offsetof
-#define offsetof(type, member) ( (int) & ((type*)0) -> member )
+#define offsetof(type, member) ( (unsigned long) & ((type*)0) -> member )
 #endif
 
 #define container_of(ptr, type, member) ({                      \
